Add files via upload
In this project I learn about the basic structure of SwiftUI apps, how to build forms and sections, creating navigation views and navigation bar titles, how to store program state with the @State property wrapper, how to create user interface controls like TextField and Picker, and how to create views in a loop using ForEach. This project is a check-sharing app that calculates how to split a check based on the number of people and how much tip you want to leave.